refactored matrix transpose tests
